+++ account/src/main/java/com/java4all/account/service/AccountService.java
@@ -12,5 +12,5 @@ public interface AccountService {
      * @param userId 用户id
      * @param money 金额
      */
-    void decrease(Integer userId, BigDecimal money);
+    void decrease(Long userId, BigDecimal money);
 }
